# feat(lua): add :verbose support for lua config