… runtime/doc/various.txt
… src/nvim/CMakeLists.txt