Currently `:verbose` doesn't show enough information when keymap/options are set from lua . It either shows 'last set from lua' or just 'last set from filename' without line no when that file is sourced . Which also becomes inacurate when you require a file that sets the option from the sourced file .

This solves that . Now proper location is  show with `:verbose`.

To get the verbose info you need tp start neovim with atleast verbose level 1 `nvim -V1` . Lua debug calls are quite slow so it's made optin.