@shadmansaleh 2021-07-13
Sorry @bfredl . This isn't ready for review yet . I wanted to run that CI so opened the pr and went to sleep . But forgot to set [WIP] in tittle. I'll let you know when it is.

Thanks for the early review by the way :D