@clason 2021-07-14
I keep forgetting which actions are set to run on draft PRs and which aren't. Maybe we can document this in Contributing.md?