@clason 2021-07-15
Commenting this here for posterity: the current failure of the `osx` action is due to https://github.com/actions/virtual-environments/issues/3733. Until this is fixed on the runner, a workaround is adding `brew remove mongodb-community` before the line https://github.com/neovim/neovim/blob/682247b52e76180e4c49b794c6cc2548bfd32d4c/.github/workflows/ci.yml#L66