@shadmansaleh 2021-07-17
I've added supoort for `vim.cmd` or `vim.api.nvim_exec` .
So if you do something like
```lua
vim.cmd 'autocmd event pattern ...'
```

In lua . Then do `:verbose au event pattern` it will give you filename and line number of `vim.cmd` 

Though this works by changing `current_sctx` in nvim_exec's c binding for lua . I'm not sure if it's the best way to do it .

Well I've added a way to have function specific code in c binding genarator . That can be reused to inject codes into lua c bindings :P