@shadmansaleh 2021-07-18
@bfredl I think it will be better if I add a global storing VIMRUNTIME . performence wise getenv all the times  is most likely bad . Also this function will be called quite often , for viml even sourcing line no is stored i  a global . If we add another one for lua it's not bad right ?

Also as far as I know VIMRUNTIME is can't change in after startup , So I can just query it once and store it at startup .

One more thing `nvim_set_option()` has `channel_id` as 1st argument it gets filled with `LUA_INTERNAL_CALL` when called from lua . That is used to set SID_LUA in that case . If I add `channel_id` to `nvim_exec` signiture will that be a breaking change to api ? I'm not sure how `channel_id` is handled for rpc calles . And confused why all api functions doesn't have `channel_id`.