@clason 2021-07-26
> Should this be merged on 5.1 or 0.6 ?

It gets merged to `master` :) 

(Which will become 0.6.0; the 0.5.1 release will come from the `release-0.5` branch, so this would have to be manually backported. This change seems to me too invasive to risk a backport for something that is a feature, not a bugfix, although if later bugfixes touch this code it might be easier to pull the commit in.)