@shadmansaleh 2021-07-26
> Which will become 0.6.0; the 0.5.1 release will come from the release-0.5 branch

Ohh I see I though master in time of 0.5.1 release will become 0.5.1 .

In that case this oviously goes to 0.6 :)