@shadmansaleh 2021-07-26
@clason benchmarks

|    Task          | Without pr |    With pr   |
|--------------------|-------------------|-----------------|
|   keymap     |       0.4 ms  |   0.92 ms  |
|   vim.opt      |    0.15 ms  |    0.67 ms  |
| Nivm_exec |    0.23 ms  |     0.77 ms |

Code : 
```lua
local now = vim.loop.hrtime
local rep = 10000

local function bench(func, kind)
  local start = now()
  for _=0, rep do
    func()
  end
  print(string.format("Each %s took on avarage %f ms", kind, (now() - start) / 1000000 / rep))
end

bench(function()
  vim.api.nvim_set_keymap('n', 'asdf', '<esc>', {})
end,"Keymap")

bench(function()
  vim.wo.number = true
end,"Option")

bench(function()
  vim.cmd [[command! TestCmd :echo "a"]]
end,"CommandOption")
```

You were rightly conserned . This is quite slow. Don't merge it now . Lets see if performence of this can be improved . But I not sure what can be done . Most of the time is spent on debug calls (lua_getinfo & lua_getstack) those are just slow .We might need a new way to get filename and line no then debug library .Not sure how else to get them though .

If anyone has an idea how to speed this up . Or how to get filename and line no without debug library please let me know .