@clason 2021-07-26
Well, it's not _quite_ as bad as I had feared :)