@bfredl 2021-07-26
I would also suggest running this benchmark on something with looks like a _real world_ init.lua, i e a long list of mappings and options (use generated mappings if needed, but then a loop that generates init.lua code, _not_ a hot loop inside the code). You are doing a hot inner loop benchmark, which will incur jit compilation of the fast path (no debug). while an init.lua is only run _once_ after parsing thus no jit speedup on toplevel statements.