@shadmansaleh 2021-07-27
@bfredl did you mean like this [bench2.lua](https://pastebin.com/qRSUrqgd)?

### Results

Sample size: 1000

|    Task   | Without pr  |    With pr   |
|-----------|-------------|--------------|
|   keymap  |    0.02 ms  |   0.047 ms   |
|   vim.opt |    0.007 ms |   0.067 ms   |
| nivm_exec |    0.016 ms |   0.049 ms   |

Ran with `build/bin/nvim -u NONE --headless +"source bench2.lua" +q`

It's somehow faster this way :P . I guess jit can't do much as it's not really lua code . 90%+ stuff it was bench marking was C .