@neovim-discourse 2021-07-27
This pull request has been mentioned on **Neovim Discourse**. There might be relevant details there:

https://neovim.discourse.group/t/how-do-i-use-space-as-the-leader-in-visual-mode/916/7