@justinmk 2021-09-22
>> optional/future: this block of code setting various fields on script_ctx looks repeated in many places throughout the codebase. Could be valuable to encapsulate it in a fuction.
>
> Good idea . Maybe in a future pr .

The more I look at this PR the more uncomfortable all of that pasted code feels. We should definitely track this as a followup.