@justinmk 2021-09-24
only blocker here imo, is the `fname_exp` question. shouldn't be too bad to preserve the old logic, and avoids risk.