@shadmansaleh 2021-09-24
> Followup work

I'll try to get them done after my exams .

Do describe these two more elaborately so I can understand what you exactly want 

> we really need to enhance :verbose for API calls. api_set_sctx(..., fn_name, code) could optionally take a fn_name and code which will say the API method name, and optionally some snippet of code which gives a hint. Also sctx should store the API channel id, so that :verbose can report the channel info (client name at least).

Do you mean it should show something like this ?
> Last set with nvim_set_keymap from channel 4